Update patch set 2
Patch Set 2:

Daniel,

I am unable to test this atm but have a question - does this charging led respect the quiet hours disable led setting? If not, it should, or we should add a way for the end user to disable the charging led altogether.  In my use case, the phone is next to my bed charging all night.  The last thing I would want is a blazing led lighting up the room. 

Another thing to consider is that the led colors can conflict with other notification light colors. In my case, yellow is gtalk, red is missed call and green is Gmail.   What would we do if a notification comes in while the phone is charging?

Patch-set: 2