Update patch set 2
Patch Set 2:

That's what i experienced tonight. It's annoying.
I hate the battery led. Useless feature.
But users cried for it.

We need a battery led menu like the notification led menu.
There's already stuff in the framework overlay.
Colors for battery level low, mid, high for example.
We would also need a option to disable it completely and a option to change the brightness, at least i9300 supports setting brightness on the led itself.

If you plugin the charger, the led starts lighting. If a notification comes in, notification values are written to the led sysfs and this overwrites the battery light. So you don't loose the notification led. After clearing the notification the battery led doesn't get retriggered. Should probably be added to the lights service.

Patch-set: 2