Update patch set 2
Patch Set 2:

I'm not sure if this is necessary with the current state of power HAL on Samsung devices.

1)  Screen-state scaling.  This is a good feature to have, however I have yet to see an implementation using the power HAL that is flexible and not hardcoded.  For example, any hardcoded power HAL settings here will conflict if the user is choosing an alternative governor.  Also, it can easily be implemented in a flexible manner using SetCPU
2)  Input performance boost - Samsung's own kernels have had this in the touchscreen driver using a DVFS lock since Gingerbread
3)  GPU stuff - since the GPU is turned off when the screen is off, screen-state scaling for GPU is pointless.

If you look at the tuna power HAL, you'll see that it offers very little beyond screen state scaling on a device that already has an input performance boost in the kernel.  The only thing that MIGHT make sense, and would require a kernel change, is something that automatically kicks the GPU up to the midrange level when the input boost is in effect.  Of course, if you now have GPU DVFS locks, may as well just add the calls for those locks in the places where CPU locks already are (assuming that a GPU lock makes sense in those places.)

Patch-set: 2