Update patch set 1
Patch Set 1:

I'm with Dan on this one. I don't really see the point of this use case. This is to enable a feature (expanded desktop), which temporarily disables another feature (navigation bar), which isn't even necessary in the first place, since there are hardware keys.

I'm not going to + or - this, since the code itself is fine. It's just my opinion that the use case is flawed.

That's just my 2 cents.

Regardless, I believe you should state dependencies on the navigation bar patches in your commit message.

Patch-set: 1