Update patch set 1
Patch Set 1:

@Kyle, there is no need to specify the navbar stuff as a dependency because it is not dependent on it - on a phone device it hides the status bar if the navbar does not exist.

We merged the Expanded desktop functionality because we believe it to have a valid use case - specifically for tablets - and, even though it has lesser value on a phone, it does not make the use case flawed.

Because we recognize that it does not have as much value on a phone as on a tablet, we ensured that the feature is overlay enabled to make it up to the device maintainers to decide if they want to enable it or not.  

My +1 is based on the fact that 1) the code is good and 2) this has been enabled on other phone devices already (tuna for example) which, if it does not make it to the i9300 is going to make our life in the XDA forums complicated as we face the inevitable 100's of questions and 'bug reports' on this 'missing' feature.

Patch-set: 1