Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I definitely understand the appeal of this future with a nav bar. I'm just not so sure the option of enabling the nav bar should be a feature in CM. I'm on the fence about that feature, but that is not part of this patch so I shouldn't discuss it here.

With that said, I didn't realize this covered the status bar as well. Thank you for the information, and please forgive my initial misunderstanding.

You've got my +1.

Patch-set: 1