Update patch set 1
Patch Set 1:

there's no option in cm to enable the navbar. there's a patch but it hasn't been merged for long. so this discussion is also useless unless the navbar patch gets merged or the i9300 looses his hardware buttons magically.

this option is one i would call "bloat" and that's what we don't want.

Patch-set: 1