Update patch set 1
Patch Set 1:

@Dan

David had specifically said NO in the below commits
http://review.cyanogenmod.com/#/c/20726/
http://review.cyanogenmod.com/#/c/20727/

There are lots of users using nav bar for i9300 as many hates the home hard key. If no one wants to approve it because "ITS USELESS" on i9300 then so be it. My own i9300 builds will have this feature enabled because it is a GREAT feature.

Patch-set: 1