Update patch set 1
Patch Set 1:

@Daniel

Assumed one is using a build with nav bar enabled.

Example of a game -> Subway surfers

Basically we just flick to jump/roll the character. Since the nav bar is present, there is a high chance that an accident flick will boot u out of the game and onto the home screen or even accidentally activating Google Now. 

I have a son who loves to play with my i9300 and he gets frustrated when he accidentally does that. With this feature, I could enlarge the screen to full size and especially important is to hide nav bar too.

It is just one example, there are many others that apps/games does NOT hide the nav bar.

Patch-set: 1