Update patch set 1
Patch Set 1:

As to, "it's useful if navbar is enabled" - Navbar is NOT enabled on this device.  If you've hacked it so it is enabled, you can make your own build with this patch.  But there is zero justification for including a feature that is only useful with unsupported hacked builds.

Patch-set: 1