Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Sorry, but this definitely the wrong approach. First reason for that is what I wrote in http://review.cyanogenmod.org/#/c/34567/. Second reason is that the dead spots created by this are way too large, as Tamas' screenshot shows.

As the S3 has virtually the same resolution as my mako (where pie works flawlessly), I suspect the problem might not be the actual pie trigger size, but the frequency of motion events. I guess the touch screen of the S3 simply sends less motion events compared to the touch screen of mako.

Jens, does that make some sense and/or have you an idea how to fix that?

Patch-set: 1