Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Yes, 15dp is definitely too much.

Currently we are using views and the fact that a gesture started in a window will be delivered completely to that window. But the downside of this is that we obscuring a small part of the screen (e.g. touch events will not be delivered to the underlying windows) and 15dp is way to much of the screen.

The real problem here is that using view attached to the window manager is more like a hack than a neat solution.

I thought about this and before I start to work on this, I want to tell it here, so you can yell at me if this is a horrible idea ;)

Luckily there is a #monitorInput() method within the InputManagerService, which is for example used by PhoneWindowManager to display the mouse pointer. Sadly this method is not part of the AIDL interface, so this is only accessible within the SystemServer realm (SystemServer.java:335 for more details). I guess was done for a good reason and I'm not going to make this part of any AIDL.

So there are two ways to get this done: 1) Build a gesture tracker within a existing system service - PhoneWindowManger comes here into mind. 2) Build a new system service that sole purpose is to expose a interface to the framework to listen for pie gesture events and when someone listens for these events, use the #monitorInput() to keep track of all the touch events.

Personally I tend to 2) since I don't want to mess around with the PhoneWindowManager beast.

Patch-set: 1