Update patch set 1
Patch Set 1:

Jens,

there's no way to propagate a touch event from one view to the view under it? That's a pity :-( Personally, I think I would go with 1) if one of the two is needed. We need to think about app expectations, though: If an app reacts on a swipe gesture from a side, and pie is enabled for that side, should the app see the swipe or not? There are arguments for either behaviour.

Patch-set: 1