Update patch set 1
Patch Set 1:

compass code is from AKM, so it should work nicely, else there is not much I can do to improve it. As for barometer, this is very surprising, since the barometer value is not supposed to change at all (unless you're on a plane I guess), so how can it be "crazy"?

Patch-set: 1