i9300: audio: properly set routing and standby after voice call
Disabling the voice call route using voicecall_default_disable is not
sufficient since there are additional routes which are used during a
voice call, which have to be properly disabled. Also, set the input
device to standy which prevents the mic from not working for the
first time after a voice call.

This could also serve a potential fix for battery drain issues after
voice call.

Change-Id: I0acd41deafd49da50857d6fe6396c3b5270d911c