i9300: libsensors: remove unused orientation sensor
The Orientation Sensor is deprecated and no longer required.
Additionally this implementation is never even used, since Android
automatically replaces it by it's own implementation if the sensor HAL
doesn't also provide a Rotation Vector sensor. Furthermore the current
implementation only uses the accelerometer and magnetometer for the
sensor fusion and doesn't make use of the Gyroscope. So should we at
some point decide to implement our own complete sensor fusion this
would have to be rewritten anyway.

Change-Id: I45d8a9afd2089b49131e6cc69cdf2f3dfee46c92