Update patch set 5
Patch Set 5:

I don't think this is acceptable at all. I thought about that before, but you're removing it from init domain which should have the permissions for that (but aosp says we should not run that on init) to run on vold, the same applies for kernel or ueventd domains. You'll still have the selinux warning about missing domain aswell

Patch-set: 5